requirements.txt
feedparser
beautifulsoup4
requests
python-dateutil
pandas
streamlit
psycopg2-binary
sqlalchemy
